Лекция 2.7
Созданы классы помощники
NavigationHelper, GroupHelper
