+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/NavigationHelper.java
@@ -1,0 +1,20 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class NavigationHelper {
+    private WebDriver wd;
+
+    public NavigationHelper(WebDriver wd) {
+        this.wd = wd;
+    }
+
+    public void gotoGroupPage() {
+        wd.findElement(By.linkText("groups")).click();
+    }
+
+    public void gotoContactPage() {
+        wd.findElement(By.linkText("add new")).click();
+    }
+}
