+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/test/ContactCreationTest.java
@@ -7,11 +7,11 @@ public class ContactCreationTest extends TestBase {
 
     @Test
     public void testContactCreation() throws Exception {
-        app.gotoContactPage();
+        app.getNavigationHelper().gotoContactPage();
         app.initContactCreation();
         app.fillContactForm(new ContactData("Ostap", "Bender", "+79998882211", "testTest@mail.ru"));
         app.submitContactCreation();
-        app.returnHomePage();
+        app.returnContactHomePage();
     }
 
 }
