+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/test/GroupCreationTests.java
@@ -7,11 +7,11 @@ public class GroupCreationTests extends TestBase {
 
     @Test
     public void testGroupCreation() throws Exception {
-        app.gotoGroupPage();
-        app.initGroupCreation();
-        app.fillGroupForm(new GroupData("Test1", "Test2", "Test3"));
-        app.submitGroupCreation();
-        app.returnGroupPage();
+        app.getNavigationHelper().gotoGroupPage();
+        app.getGroupHelper().initGroupCreation();
+        app.getGroupHelper().fillGroupForm(new GroupData("Test1", "Test2", "Test3"));
+        app.getGroupHelper().submitGroupCreation();
+        app.getGroupHelper().returnGroupPage();
     }
 
 }
