+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/test/GroupDeletionTest.java
@@ -7,10 +7,10 @@ public class GroupDeletionTest extends TestBase {
 
     @Test
     public void testGroupDeletion() throws Exception {
-        app.gotoGroupPage();
-        app.selectGroup();
-        app.deleteSelectedGroup();
-        app.returnGroupPage();
+        app.getNavigationHelper().gotoGroupPage();
+        app.getGroupHelper().selectGroup();
+        app.getGroupHelper().deleteSelectedGroup();
+        app.getGroupHelper().returnGroupPage();
     }
 
 }
